+++ upload/library/Iversia/FAQ/AttachmentHandler/Question.php
@@ -2,27 +2,50 @@
 
 class Iversia_FAQ_AttachmentHandler_Question extends XenForo_AttachmentHandler_Abstract
 {
-    protected $_contentRoute           = 'faq';
-    protected $_contentTypePhraseKey   = 'faq';
-    protected $_contentIdKey           = 'faq_id';
+    protected $_contentIdKey = 'faq_id';
+    protected $_contentRoute = 'faq';
+    protected $_contentTypePhraseKey = 'xf_faq_question';
 
     protected function _canUploadAndManageAttachments(array $contentData, array $viewingUser)
     {
-        return true;
+        return ($viewingUser['user_id']
+            && XenForo_Permission::hasPermission($viewingUser['permissions'], 'FAQ_Manager_Permissions', 'uploadFAQAttach')
+        );
     }
 
-    public function _canViewAttachment(array $attachment, array $viewingUser)
+    // Everyone can view attachments on FAQ entries
+    protected function _canViewAttachment(array $attachment, array $viewingUser)
     {
         return true;
     }
 
+    // What to do after an attachment is deleted
     public function attachmentPostDelete(array $attachment, Zend_Db_Adapter_Abstract $db)
     {
+        $db->query('
+            UPDATE xf_faq_question
+            SET attach_count = IF(attach_count > 0, attach_count - 1, 0)
+            WHERE faq_id = ?
+        ', $attachment['content_id']);
+
         return true;
     }
 
-    public function getAttachmentCountLimit()
+    protected function _getQuestionModel()
     {
-        return true;
+        return XenForo_Model::create('Iversia_FAQ_Model_Question');
     }
+
+ /*
+
+
+    public function getContentLink(array $attachment, array $extraParams = array(), $skipPrepend = false)
+    {
+        $question = $this->getContentDataFromContentId($attachment['content_id']);
+        $extraParams['faq_id'] = $question['faq_id'];
+        $extraParams['question'] = $question['question'];
+        return XenForo_Link::buildPublicLink($this->_contentRoute, $update, $extraParams, $skipPrepend);
+    }
+
+    */
 }