+++ upload/library/Iversia/FAQ/DataWriter/Question.php
@@ -2,6 +2,10 @@
 
 class Iversia_FAQ_DataWriter_Question extends XenForo_DataWriter
 {
+    protected $_question = null;
+
+    const DATA_ATTACHMENT_HASH = 'attachmentHash';
+
     protected function _getFields()
     {
         return array(
@@ -15,6 +19,7 @@ class Iversia_FAQ_DataWriter_Question extends XenForo_DataWriter
 
                 'user_id'       => array('type' => self::TYPE_UINT,     'required' => true),
 
+                'attach_count'  => array('type' => self::TYPE_UINT,     'required' => false, 'default' => 0),
                 'view_count'    => array('type' => self::TYPE_UINT,     'required' => false),
                 'likes'         => array('type' => self::TYPE_UINT,     'required' => false, 'default' => 0),
                 'like_users'    => array('type' => self::TYPE_SERIALIZED, 'default' => 'a:0:{}'),
@@ -46,11 +51,46 @@ class Iversia_FAQ_DataWriter_Question extends XenForo_DataWriter
     protected function _postSave()
     {
         $this->_indexForSearch();
+
+        $attachmentHash = $this->getExtraData(self::DATA_ATTACHMENT_HASH);
+        if ($attachmentHash)
+        {
+            $this->_associateAttachments($attachmentHash);
+        }
+    }
+
+    protected function _deleteAttachments()
+    {
+        $this->getModelFromCache('XenForo_Model_Attachment')->deleteAttachmentsFromContentIds(
+            'xf_faq_question',
+            array($this->get('faq_id'))
+        );
     }
 
     protected function _postDelete()
     {
         $this->_deleteFromSearchIndex();
+        $this->_deleteAttachments();
+    }
+
+    protected function _associateAttachments($attachmentHash)
+    {
+        $rows = $this->_db->update('xf_attachment', array(
+            'content_type' => 'xf_faq_question',
+            'content_id' => $this->get('faq_id'),
+            'temp_hash' => '',
+            'unassociated' => 0
+        ), 'temp_hash = ' . $this->_db->quote($attachmentHash));
+
+        if ($rows) {
+            $newAttachCount = $this->get('attach_count') + $rows;
+
+            $this->set('attach_count', $newAttachCount, '', array('setAfterPreSave' => true));
+            $this->_db->update('xf_faq_question', array(
+                'attach_count' => $newAttachCount),
+                'faq_id = ' . $this->get('faq_id')
+            );
+        }
     }
 
     protected function _indexForSearch()
@@ -80,5 +120,10 @@ class Iversia_FAQ_DataWriter_Question extends XenForo_DataWriter
     {
         // Gets the search-data handler for 'xf_faq_question' content type
         return $this->getModelFromCache('XenForo_Model_Search')->getSearchDataHandler('xf_faq_question');
+    }
+
+    public function setQuestion(array $question)
+    {
+        $this->_question = $question;
     }
 }