+++ upload/library/Iversia/FAQ/Listener/Extend.php
@@ -28,6 +28,6 @@ class Iversia_FAQ_Listener_Extend
     {
         return $matches[0] .
         '<xen:if is="{$controllerName} == \'Iversia_FAQ_ControllerPublic_FAQ\'">
-        <div id="iversiaFAQ" class="muted">FAQ Manager &copy;2013 <a href="http://www.iversia.com" title="Iversia - Web Developer and Illustrator" class="concealed">Iversia</a> from <a href="http://www.rpgfix.com" title="RPGfix, Role-Playing Game Directory" class="concealed">RPGfix</a>.</div></xen:if>';
+        <div id="iversiaFAQ" class="muted">FAQ Manager &copy;'.date('Y').' <a href="http://www.iversia.com" title="Iversia - Web Developer and Illustrator" class="concealed">Iversia</a> from <a href="http://www.rpgfix.com" title="RPGfix, Role-Playing Game Directory" class="concealed">RPGfix</a>.</div></xen:if>';
     }
 }