+++ upload/library/Iversia/FAQ/ViewPublic/Index.php
@@ -7,12 +7,17 @@ class Iversia_FAQ_ViewPublic_Index extends XenForo_ViewPublic_Base
         $bbCodeParser = new XenForo_BbCode_Parser(XenForo_BbCode_Formatter_Base::create('Base', array('view' => $this)));
         $bbCodeOptions = array(
             'states' => array(
-                'viewAttachments' => false
-            )
+                'viewAttachments' => true
+            ),
+            'contentType' => 'xf_faq_question',
+            'contentIdKey' => 'faq_id',
+            'messageKey' => 'answer',
         );
 
         foreach ($this->_params['faq'] as &$question) {
-            $question['answer'] = new XenForo_BbCode_TextWrapper($question['answer'], $bbCodeParser);
+
+            $question['answer'] = XenForo_ViewPublic_Helper_Message::getBbCodeWrapper($question, $bbCodeParser, $bbCodeOptions);
+
         }
     }
 }